refactor : map API 따로 부리
