+++ src/map/map.module.ts
@@ -1,0 +1,10 @@
+// Map.module.ts
+import { Module } from '@nestjs/common';
+import { MapController } from './map.controller';
+import { MapService } from './map.service';
+
+@Module({
+  controllers: [MapController],
+  providers: [MapService],
+})
+export class MapModule {}
